feat(konnect): set KongConsumer conditions when specified credentials are missing
